Move Provider namespace into Bridge\Silex
The Silex1 was added to the provider class in order to prepare Silex 2.0 compatibility.
