+++ UPGRADE-2.1.md
@@ -5,3 +5,9 @@
 The `SLLH\IsoCodesValidator\Bundle` namespace is deprecated.
 
 Use `SLLH\IsoCodesValidator\Bridge\Symfony` namespace instead.
+
+## Deprecated Provider namespace
+
+The `SLLH\IsoCodesValidator\Provider` namespace is deprecated.
+
+Use `SLLH\IsoCodesValidator\Bridge\Silex` namespace instead.
