+++ phpunit.xml.dist
@@ -23,7 +23,7 @@
         <testsuite name="silex">
             <directory>./tests/Constraints</directory>
             <file>./tests/AbstractIsoCodeGenericConstraintTest.php</file>
-            <directory>./tests/Provider</directory>
+            <directory>./tests/Bridge/Silex</directory>
         </testsuite>
     </testsuites>
