+++ src/Bridge/Silex/IsoCodesValidatorSilex1ServiceProvider.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace SLLH\IsoCodesValidator\Bridge\Silex;
+
+use Silex\Application;
+use Silex\ServiceProviderInterface;
+
+/**
+ * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ */
+class IsoCodesValidatorSilex1ServiceProvider implements ServiceProviderInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function register(Application $app)
+    {
+        if (isset($app['translator'])) {
+            $file = __DIR__.'/../../Resources/translations/validators.'.$app['locale'].'.xlf';
+            if (file_exists($file)) {
+                $app['translator']->addResource('xliff', $file, $app['locale'], 'validators');
+            }
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function boot(Application $app)
+    {
+    }
+}
