+++ src/Provider/IsoCodesValidatorServiceProvider.php
@@ -2,31 +2,19 @@
 
 namespace SLLH\IsoCodesValidator\Provider;
 
-use Silex\Application;
-use Silex\ServiceProviderInterface;
+use SLLH\IsoCodesValidator\Bridge\Silex\IsoCodesValidatorSilex1ServiceProvider;
+
+@trigger_error(
+    'The '.__NAMESPACE__.'\IsoCodesValidatorServiceProvider class is deprecated since version 3.1 and will be removed in 4.0.'
+    .' Use SLLH\IsoCodesValidator\Bridge\Silex\IsoCodesValidatorSilex1ServiceProvider.',
+    E_USER_DEPRECATED
+);
 
 /**
  * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ *
+ * @deprecated since version 2.1, to be removed in 3.0. Use SLLH\IsoCodesValidator\Bridge\Silex\IsoCodesValidatorSilex1ServiceProvider instead.
  */
-class IsoCodesValidatorServiceProvider implements ServiceProviderInterface
+class IsoCodesValidatorServiceProvider extends IsoCodesValidatorSilex1ServiceProvider
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function register(Application $app)
-    {
-        if (isset($app['translator'])) {
-            $file = __DIR__.'/../Resources/translations/validators.'.$app['locale'].'.xlf';
-            if (file_exists($file)) {
-                $app['translator']->addResource('xliff', $file, $app['locale'], 'validators');
-            }
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function boot(Application $app)
-    {
-    }
 }
